example install script
I like to keep notes of my install in the form of a script.
This is just the debian version, (but I usually add a `yum` version
as well.)

The idea is to make install effortless, in the form of:

`curl --proto '=https' --tlsv1.2 -sSfL https://raw.githubusercontent.com/alexxroche/delta/master/install_revolt_server.sh | sh`

to improve adoption.