chore: bump deps
Not sure what to do in wait_for_event() in test.rs when the item returns a RedisError. For now opted to just unwrap() since this is a test.

jxl_oxide has deprecated Render::image() in favour of Render::stream()

Comment out unused redis22

In Rocket, Outcome::Failure has become Outcome::Error

to_rfc3339_string() is deprecated in favour of it's equivalent: try_to_rfc3339_string().unwrap()
This should be ok because we are calling this method on DateTime::now()

Signed-off-by: Pi-Cla <pirateclip@protonmail.com>